backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg2==2.9.10
djangorestframework-simplejwt==5.3.1
django-clone==5.3.3
django-filter==24.3